flashtool: align flash program start address to start of block, with padding
The i.MX RAM kernel erases before it programs automatically, so we do not need
to explicity erase.  If the user requests to start writing the flash file
to an address that is not block-aligned, pre-pad the data so it can be
block-aligned.

This may be useful to add as a utility method in the RKL library, but for now
leave the logic in flashtool.