pyatk: fix "errata" with i.MX25 USB and SerialBootProtocol.write_file() loading files on a 64-byte boundary
It seems that if we write_file(FILE_TYPE_APPLICATION) with a stream that is a multiple of 64 bytes, the complete_boot()
command would not complete if only 16 bytes were sent.  If we kick an additional byte before calling complete_boot(),
we can avoid the issue entirely (and it does NOT affect images that do not fall on a 64-byte boundary, at least on the
i.MX25).

Also internalize complete_boot(), and call it automatically when write_file(FILE_TYPE_APPLICATION) is used.